parser/requirements.txt
requests
fastapi
uvicorn
aiohttp
beautifulsoup4
sqlalchemy
sqlmodel
psycopg2
python-dotenv
celery[redis]